refactor: 마이페이지 관련 기능 멤버에서 분할
