+++ src/main/java/BE_Elixir/Elixir/domain/member/controller/MemberController.java
@@ -7,8 +7,6 @@ import BE_Elixir.Elixir.domain.member.dto.response.*;
 import BE_Elixir.Elixir.domain.member.entity.Member;
 import BE_Elixir.Elixir.domain.member.entity.MemberDetails;
 import BE_Elixir.Elixir.domain.member.service.MemberService;
-import BE_Elixir.Elixir.domain.recipe.dto.response.RecipeImageResponseDTO;
-import BE_Elixir.Elixir.global.exception.ErrorCode;
 import BE_Elixir.Elixir.global.redis.RedisAuthService;
 import BE_Elixir.Elixir.global.response.CommonResponse;
 import BE_Elixir.Elixir.global.security.JwtProvider;
@@ -123,116 +121,6 @@ public class MemberController implements MemberApi {
         return ResponseEntity.ok(CommonResponse.success(HttpStatus.OK.value(), HttpStatus.OK.toString(), "회원탈퇴 성공"));
     }
 
-    // 회원 정보 조회 (이메일, 닉네임, 젠더, 생년, 프로필 url)
-    @GetMapping()
-    public ResponseEntity<CommonResponse<MemberResponseDTO>> getMemberInfo(
-            @AuthenticationPrincipal MemberDetails memberDetails,
-            HttpServletRequest request)
-    {
-        String email = memberDetails.getUsername();
-        log.info("회원 정보 조회 요청 - email: {}", email);
-
-        MemberResponseDTO response = memberService.getMemberInfo(email);
-        log.info("회원 정보 조회 성공 - email: {}", email);
-
-        return ResponseEntity.ok(CommonResponse.success(HttpStatus.OK.value(), HttpStatus.OK.toString(), "회원조회 성공", response));
-    }
-
-    // 로그인한 사용자 프로필 수정 시, 얻은 칭호 목록 조회
-    @GetMapping("/achievement/title")
-    public ResponseEntity<CommonResponse<MemberTitlesResponseDTO>> getTitles(
-            @AuthenticationPrincipal MemberDetails memberDetails
-    ) {
-        Long memberId = memberDetails.getId();
-        log.info("칭호 목록 조회 요청 - id: {}", memberId);
-
-        List<String> titles = memberService.getTitles(memberId);
-        MemberTitlesResponseDTO dto = MemberTitlesResponseDTO.builder()
-                .memberId(memberId)
-                .titles(titles)
-                .build();
-
-        log.info("칭호 목록 조회 성공 - id: {}", memberId);
-
-        return ResponseEntity.ok(CommonResponse.success(HttpStatus.OK.value(), HttpStatus.OK.toString(), "칭호 목록 성공", dto));
-    }
-
-    // 로그인한 사용자 프로필 수정하기
-    @PatchMapping(value="/profile", consumes = MediaType.MULTIPART_FORM_DATA_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<CommonResponse<MemberResponseDTO>> updateMemberProfile(
-            @RequestPart(value = "dto", required = false) MemberProfileRequestDTO dto,
-            @RequestPart(value = "profileImage", required = false) MultipartFile profileImage,
-            @AuthenticationPrincipal MemberDetails memberDetails
-    ) {
-        Long memberId = memberDetails.getId();
-        log.info("사용자 프로필 수정 요청 - id: {}", memberId);
-
-        MemberResponseDTO responseDTO = memberService.updateMemberProfile(memberId, dto, profileImage);
-
-        log.info("사용자 프로필 수정 성공 - id: {}", memberId);
-
-        return ResponseEntity.ok(CommonResponse.success(HttpStatus.OK.value(), HttpStatus.OK.toString(), "사용자 프로필 수정 성공", responseDTO));
-    }
-
-    // 로그인한 사용자 프로필 조회 (칭호, 닉네임, 프로필사진, 팔로워 수, 팔로잉 수)
-    @GetMapping("/profile")
-    public ResponseEntity<CommonResponse<MemberProfileResponseDTO>> getMemberProfile(
-            @AuthenticationPrincipal MemberDetails memberDetails
-    ) {
-        Long memberId = memberDetails.getId();
-        log.info("로그인한 사용자의 프로필 조회 요청 - id: {}", memberId);
-
-        MemberProfileResponseDTO responseDTO = memberService.getMemberProfile(memberId);
-
-        log.info("로그인한 사용자의 프로필 조회 성공 - id: {}", memberId);
-
-        return ResponseEntity.ok(CommonResponse.success(HttpStatus.OK.value(), HttpStatus.OK.toString(), "로그인한 사용자의 프로필 조회 성공", responseDTO));
-    }
-
-
-    // 특정 사용자 프로필 조회 (칭호, 닉네임, 프로필사진, 팔로워 수, 팔로잉 수)
-    @GetMapping("/{memberId}/profile")
-    public ResponseEntity<CommonResponse<MemberProfileResponseDTO>> getOtherMemberProfile(
-            @PathVariable("memberId") Long memberId,
-            @AuthenticationPrincipal MemberDetails memberDetails
-    ) {
-        log.info("특정 사용자의 프로필 조회 요청 - id: {}", memberId);
-
-        MemberProfileResponseDTO responseDTO = memberService.getMemberProfile(memberId);
-
-        log.info("특정 사용자의 프로필 조회 성공 - id: {}", memberId);
-
-        return ResponseEntity.ok(CommonResponse.success(HttpStatus.OK.value(), HttpStatus.OK.toString(), "로그인한 사용자의 프로필 조회 성공", responseDTO));
-    }
-
-    // 로그인한 사용자가 업로드한 모든 레시피 조회하기
-    @GetMapping("/recipe")
-    public ResponseEntity<CommonResponse<List<RecipeImageResponseDTO>>> getMyRecipes(
-            @AuthenticationPrincipal MemberDetails memberDetails
-    ) {
-        Long memberId = memberDetails.getId();
-
-        List<RecipeImageResponseDTO> recipes = memberService.getMyRecipes(memberId);
-        return ResponseEntity.ok(CommonResponse.success(
-                HttpStatus.OK.value(), HttpStatus.OK.toString(), "내 레시피 조회 성공", recipes));
-
-    }
-
-
-    // 로그인한 사용자가 스크랩한 레시피 조회하기
-    @GetMapping("/recipe/scrap")
-    public ResponseEntity<CommonResponse<List<RecipeImageResponseDTO>>> getMyScrapRecipes(
-            @AuthenticationPrincipal MemberDetails memberDetails
-    ) {
-        Long memberId = memberDetails.getId();
-
-        List<RecipeImageResponseDTO> scrappedRecipes = memberService.getMyScrapRecipes(memberId);
-
-        return ResponseEntity.ok(CommonResponse.success(
-                HttpStatus.OK.value(), HttpStatus.OK.toString(),
-                "내가 스크랩한 레시피 조회 성공", scrappedRecipes));
-    }
-
     // 팔로우 하기
     @PostMapping("/{targetMemberId}/follow")
     public ResponseEntity<CommonResponse<?>> follow(
@@ -291,7 +179,6 @@ public class MemberController implements MemberApi {
                 "현 사용자의 팔로워 목록 조회 성공", dto));
     }
 
-
     // (특정 사용자의) 팔로잉 목록 조회하기 (사용자가 팔로우하는 목록)
     @GetMapping("/{targetMemberId}/following")
     public ResponseEntity<CommonResponse<List<MemberSummaryDTO>>> getFollowingByMemberId(
@@ -304,7 +191,6 @@ public class MemberController implements MemberApi {
                 "특정 사용자의 팔로잉 목록 조회 성공", dto));
     }
 
-
     // (특정 사용자의) 팔로우 목록 조회하기 (사용자를 팔로잉하는 목록)
     @GetMapping("/{targetMemberId}/follower")
     public ResponseEntity<CommonResponse<List<MemberSummaryDTO>>> getFollowerByMemberId(
@@ -315,95 +201,5 @@ public class MemberController implements MemberApi {
         return ResponseEntity.ok(CommonResponse.success(
                 HttpStatus.OK.value(), HttpStatus.OK.toString(),
                 "특정 사용자의 팔로워 목록 조회 성공", dto));
-    }
-
-    // 로그인한 사용자의 모든 챌린지 업적 정보 조회
-    @GetMapping("/achievement")
-    public ResponseEntity<CommonResponse<List<MemberAchievementResponseDTO>>> getAllAchievements(
-            @AuthenticationPrincipal MemberDetails memberDetails
-    ) {
-        Long memberId = memberDetails.getId();
-
-        List<MemberAchievementResponseDTO> achievements = memberService.getAllAchievements(memberId);
-
-        return ResponseEntity.ok(CommonResponse.success(
-                HttpStatus.OK.value(), HttpStatus.OK.toString(),
-                "모든 챌린지 업적 조회 성공", achievements));
-    }
-
-    // 로그인한 사용자의 달성한 업적 최신 3개 조회하기
-    @GetMapping("/achievement/top3")
-    public ResponseEntity<CommonResponse<List<MemberAchievementResponseDTO>>> getTop3Achievements(
-            @AuthenticationPrincipal MemberDetails memberDetails
-    ) {
-        Long memberId = memberDetails.getId();
-
-        List<MemberAchievementResponseDTO> top3Achievements = memberService.getTop3Achievements(memberId);
-
-        return ResponseEntity.ok(CommonResponse.success(
-                HttpStatus.OK.value(), HttpStatus.OK.toString(),
-                "최근 업적 3개 조회 성공", top3Achievements));
-    }
-
-    // 로그인한 사용자의 설문조사 결과 조회하기
-    @GetMapping("/survey")
-    public ResponseEntity<CommonResponse<SurveyResponseDTO>> getSurvey(
-            @AuthenticationPrincipal MemberDetails memberDetails
-    ) {
-        Long memberId = memberDetails.getId();
-
-        SurveyResponseDTO responseDTO = memberService.getSurvey(memberId);
-
-        return ResponseEntity.ok(CommonResponse.success(
-                HttpStatus.OK.value(), HttpStatus.OK.toString(),
-                "설문조사 결과 조회 성공, 회원 ID: " + memberId, responseDTO));
-    }
-
-    // 로그인한 사용자의 설문조사 결과 수정하기
-    @PutMapping("/survey")
-    public ResponseEntity<CommonResponse<SurveyResponseDTO>> updateSurvey(
-            @RequestBody SurveyRequestDTO dto,
-            @AuthenticationPrincipal MemberDetails memberDetails
-    ) {
-        Long memberId = memberDetails.getId();
-
-        SurveyResponseDTO responseDTO = memberService.updateSurvey(memberId, dto);
-
-        return ResponseEntity.ok(CommonResponse.success(
-                HttpStatus.OK.value(), HttpStatus.OK.toString(),
-                "설문조사 수정 성공, 회원 ID: " + memberId, responseDTO));
-    }
-
-    // 다른 사용자가 업로드한 모든 레시피 조회하기
-    @GetMapping("/{memberId}/recipes")
-    public ResponseEntity<CommonResponse<List<RecipeImageResponseDTO>>> getUserRecipes(
-            @PathVariable("memberId") Long memberId
-    ) {
-        List<RecipeImageResponseDTO> recipes = memberService.getUserRecipes(memberId);
-        return ResponseEntity.ok(CommonResponse.success(
-                HttpStatus.OK.value(), HttpStatus.OK.toString(), "다른 사용자 레시피 조회 성공", recipes));
-
-    }
-
-    // 다른 사용자의 모든 챌린지 업적 정보 조회하기
-    @GetMapping("/{memberId}/achievements")
-    public ResponseEntity<CommonResponse<List<MemberAchievementResponseDTO>>> getAllAchievementsByMemberId(
-            @PathVariable Long memberId
-    ) {
-        List<MemberAchievementResponseDTO> achievements = memberService.getAllAchievementsByMemberId(memberId);
-        return ResponseEntity.ok(CommonResponse.success(
-                HttpStatus.OK.value(), HttpStatus.OK.toString(),
-                "다른 사용자의 모든 업적 조회 성공", achievements));
-    }
-
-    // 다른 사용자의 최근 3개 업적 조회하기
-    @GetMapping("/{memberId}/achievements/top3")
-    public ResponseEntity<CommonResponse<List<MemberAchievementResponseDTO>>> getTop3AchievementsByMemberId(
-            @PathVariable Long memberId
-    ) {
-        List<MemberAchievementResponseDTO> achievements = memberService.getTop3AchievementsByMemberId(memberId);
-        return ResponseEntity.ok(CommonResponse.success(
-                HttpStatus.OK.value(), HttpStatus.OK.toString(),
-                "다른 사용자의 최근 업적 3개 조회 성공", achievements));
     }
 }
